Use atomic writes for copy or move of local files
If multiple processes are copying to the same destination
we want the content of the file to be correct even if we
don't know which file we end up with.

Note that os.rename does not care if the file already
exists so the overwrite parameter is ignored if the
target file appeared between the first check and the
final rename.