Restructure tests and remove testing with mocked responses.
We are using an external webDAV server to test against. In addition
we have the possibility now to test against other real webDAV servers
in the develpment setting by specifying via the environment variable
LSST_RESOURCES_HTTP_TEST_SERVER_URL the URL of a server to test
against.

However, since we are no longer mocking responses, in the current
situation we cannot fully test cases where the server does not
support some features (e.g. a plain HTTP server, or a webDAV server
that does not support download a specified range of bytes).