Persist connections to server but ensure that connections open by the parent process are reset in the child process.
In addition, we ask the server to compute a digest at upload time.
The digest is not computed by the client so we don't check that
the checksum computed by the server matches ours, but it seems
useful to ask the server, if it implements this, to record
each file's checksum.