Fix running off the end of files over http and s3
Both s3 and http file handles use byte ranges to request subsequent
byte reads. Fix an issue where the requested range is past the end
of the file by catching response codes and requesting the rest of
the file where appropriate. The handles will return empty byte
string when there is nothing remaining in the file.