Disable gzip encoding from server when using byte range
If the server supports gzip encoding (which is the default for
urllib) the byte ranges refer to bytes in the compressed version
of the content. This can break the client because there is no
guarantee that those bytes can be uncompressed. Instead disable
the Accept-Encoding header so that the server is forced to
use the original byte range.