Fix issue with restrictive umasks
When creating directories, we now modify the filesystem permissions to
add the write and execute bits for the current user if they aren't
already set.  This fixes exceptions caused by filesystem permissions
when attempting to create files in read-only directories. This issue was
occurring because users had set the global process umask to 0222 in an
attempt to force files to be read-only.