Fix HttpResourcePath.write() for S3 URLs
Limit the parent directory creation in HttpResourcePath.write() to WebDAV endpoints only.  Presigned S3 URLs do not have any way to check for existence of the parent directory.  S3 has no concept of directories and even if it did, you wouldn't be able to do anything about them via a URL signed for the child.