No longer call isdir() on schemeless URI
This would only work if the file pointed to by the schemeless URI
happened to exist. Given that a schemeless URI could be relative
to any location, we should not be checking relative to cwd unless
we are really asked to see if this is a directory or to read it.