Modify how dirLike is calculated
For schemeless URIs we do not know whether something is a directory
or not without checking the file system but we should not do that
until we are asked to find out. Add a None state to the dirLike
property to indicate that directory status is not yet known.
This is only relevant for files since S3 resources rely on
the trailing slash.