Fix memory leak in HTTPResourcePath
Fixed an issue where query parameters or fragments in an HTTP URL would cause it to be considered as having a different root URI.  This caused a new HTTP session to be created and cached each time an HTTP URL with a different query parameter was accessed.

This resulted in a memory leak in datalinker when it was using HTTPResourcePath to access S3 presigned URLs, which always have unique query parameters.