Change default directory used for temp files
Previously, if there was not an environment variable set explicitly defining the location of the temporary directory, HttpResourcePath would use the current working directory during downloads.  This is a bad default for a few reasons:
1. It has frequently caused issues when using the library on systems with read-only directories (like containers and notebooks.)
2. On many systems where the library is used, the current working directory can be a slow network filesystem.
3.  It's surprising in general -- all operating systems have a well-defined default temporary directory.

Instead we now use the same directory as Python's default temporary directory, which chooses something appropriate based on the environment and operating system.