Force a default AWS region for S3 tests
This prevents problems when there is a local override in
~/.aws/config.

Note that we use "us-east-1" for all tests. "us-west-1" does
not work. In theory MOTO_ALLOW_NONEXISTENT_REGION could be
set to allow an arbitrary fake region to be used but that
does not seem to work with our tests at present.