Propagate fragments with ResourcePath.join
Previously fragments were either dropped completely (if a
ResourcePath was given that had a fragment) or included in the
file name (if a plain string was given).

Now fragments are always propagated if the ResourcePath parsing
finds a fragment. This means that if a filename has a "#" in it
that "#" will now always be treated as a fragment and not part
of the filename.